services/engine/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Config & validation
pydantic-settings==2.7.0
python-dotenv==1.0.1

# HTTP client (crawler)
httpx==0.28.1

# HTML parsing (crawler)
beautifulsoup4==4.12.3
lxml==5.3.0

# Database (async, for Fase 2b+)
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# Redis (job queue, for Fase 2b+)
redis==5.2.1

# AI / LLM (Fase 3+)
anthropic==0.42.0

# Embeddings (Fase 3+) — ONNX runtime, no PyTorch required (~200 MB Docker layer)
fastembed==0.4.2
